pyproject.toml
[tool.poetry]
name = "finance-app"
version = "0.0.0"
description = "finance-app"
authors = ["pavivin <pavivin@yandex.ru>"]
readme = "README.md"
packages = [{include = "finance-app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.86.0"
orjson = "^3.8.1"
aiobotocore = "^2.4.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
python-multipart = "^0.0.5"
redis = "^4.3.4"
websockets = "^10.4"
uvicorn = {extras = ["standart"], version = "^0.19.0"}
yfinance = "^0.1.87"
googlefinance = "^0.7"
wallstreet = "^0.3.1"
setuptools = "^65.5.1"
realtime-stock = "^1.0.0"
python-socketio = "^5.7.2"
yahoo-finance = "^1.4.0"


[tool.black]
line-length = 120
target-version = ['py310', 'py311']
skip-string-normalization = true
include = '\.pyi?$'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
